feat: 고정바우처, 퍼센트바우처 제거 후, 계산하는 Discount클래스 합성
