+++ src/main/java/org/weekly/weekly/voucher/domain/Discount.java
@@ -1,0 +1,4 @@
+package org.weekly.weekly.voucher.domain;
+
+public interface Discount {
+}
