+++ src/main/java/org/weekly/weekly/voucher/domain/FixedDiscount.java
@@ -1,0 +1,4 @@
+package org.weekly.weekly.voucher.domain;
+
+public class FixedDiscount implements Discount{
+}
