--- src/main/java/org/weekly/weekly/voucher/domain/FixedDiscountVoucher.java
@@ -1,24 +1,0 @@
-package org.weekly.weekly.voucher.domain;
-
-import java.time.LocalDate;
-import java.util.UUID;
-
-public class FixedDiscountVoucher implements Voucher {
-    private final UUID voucherId;
-    private final long amount;
-
-    private final LocalDate registrationDate;
-    private final LocalDate expirationDate;
-
-    public FixedDiscountVoucher(UUID voucherId, long amount, LocalDate registrationDate, LocalDate expirationDate) {
-        this.voucherId = voucherId;
-        this.amount = amount;
-        this.registrationDate = registrationDate;
-        this.expirationDate = expirationDate;
-    }
-    
-    @Override
-    public UUID getVoucherId() {
-        return voucherId;
-    }
-}
