+++ src/main/java/org/weekly/weekly/voucher/domain/PercentDiscount.java
@@ -1,0 +1,4 @@
+package org.weekly.weekly.voucher.domain;
+
+public class PercentDiscount implements Discount{
+}
